+++ playground/knot/src/main.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import Chrome from "./Editor/Chrome";
+import Playground from "./Playground";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <Chrome />
+    <Playground />
   </React.StrictMode>,
 );
